@@ -1,35 +1,40 @@
 {
-  "name": "redux-immutable-todo",
   "private": true,
+  "scripts": {
+    "lint": "eslint ./src",
+    "start": "webpack-dev-server",
+    "precommit": "npm run lint"
+  },
   "devDependencies": {
-    "babel": "^5.8.23",
-    "babel-eslint": "^4.1.0",
-    "babel-loader": "^5.3.2",
-    "chai": "^3.2.0",
-    "eslint": "^1.2.1",
-    "eslint-plugin-react": "^3.3.0",
-    "gulp": "^3.9.0",
-    "gulp-eslint": "^1.0.0",
-    "gulp-mocha": "^2.1.3",
-    "mocha": "^2.2.5",
-    "react-addons-test-utils": "^0.14.0-beta3",
-    "react-hot-loader": "^1.2.9",
-    "redux-devtools": "^1.1.2",
-    "sinon": "^1.16.1",
-    "style-loader": "^0.12.3",
-    "webpack": "^1.12.0",
-    "webpack-dev-server": "^1.10.1"
+    "babel": "^6.5.2",
+    "babel-eslint": "^7.1.1",
+    "babel-loader": "^6.2.10",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.16.0",
+    "babel-preset-stage-0": "^6.16.0",
+    "chai": "^3.5.0",
+    "css-loader": "^0.26.1",
+    "eslint": "^3.13.1",
+    "eslint-config-canonical": "^6.1.0",
+    "react-hot-loader": "^1.3.1",
+    "redux-devtools": "^3.3.1",
+    "style-loader": "^0.13.1",
+    "webpack": "^1.14.0",
+    "webpack-dev-server": "^1.16.2"
   },
   "dependencies": {
-    "classnames": "^2.1.3",
-    "immutable": "^3.7.4",
-    "lodash": "^3.10.1",
-    "react": "^0.14.0-beta3",
-    "react-dom": "^0.14.0-beta3",
-    "react-redux": "^1.0.1",
-    "redux": "^1.0.1",
-    "redux-immutable": "^1.3.2",
-    "redux-thunk": "^0.1.0",
-    "reselect": "^0.0.2"
+    "babel-core": "^6.21.0",
+    "classnames": "^2.2.5",
+    "immutable": "^3.8.1",
+    "lodash": "^4.17.4",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-redux": "^5.0.2",
+    "redux": "^3.6.0",
+    "redux-create-reducer": "^1.1.1",
+    "redux-immutable": "^3.0.9",
+    "redux-logger": "^2.7.4",
+    "redux-thunk": "^2.1.0",
+    "reselect": "^2.5.4"
   }
 }
