@@ -1,15 +1,13 @@
 /* eslint-disable no-unused-vars */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './containers/App';
-
 import {
-    Provider
+  Provider
 } from 'react-redux';
-
+import App from './containers/App';
 import store from './store';
 
 ReactDOM.render(
-    <Provider store={store}>
-        {() => <App />}
-    </Provider>, document.getElementById('app'));
+  <Provider store={store}>
+    <App />
+  </Provider>, document.getElementById('app'));
