@@ -1,90 +1,84 @@
 import _ from 'lodash';
-
+import {
+  createReducer
+} from 'redux-create-reducer';
 import Immutable from 'immutable';
 
-let CONSTRUCT,
-    TASK_ADD,
-    TASK_DONE,
-    TASK_UNDONE;
+const initialState = Immutable.fromJS([
+  {
+    done: true,
+    id: _.uniqueId(),
+    name: 'foo'
+  },
+  {
+    done: false,
+    id: _.uniqueId(),
+    name: 'bar'
+  },
+  {
+    done: false,
+    id: _.uniqueId(),
+    name: 'baz'
+  },
+  {
+    done: false,
+    id: _.uniqueId(),
+    name: 'quux'
+  }
+]);
 
-CONSTRUCT = () => {
-    return Immutable.fromJS([
-        {
-            id: _.uniqueId(),
-            name: 'foo',
-            done: true
-        },
-        {
-            id: _.uniqueId(),
-            name: 'bar',
-            done: false
-        },
-        {
-            id: _.uniqueId(),
-            name: 'baz',
-            done: false
-        },
-        {
-            id: _.uniqueId(),
-            name: 'quux',
-            done: false
-        }
-    ]);
+/**
+ * @param {Immutable.List} domain
+ * @param {Object} action
+ * @param {string} action.data.name
+ * @returns {Immutable.List}
+ */
+const TASK_ADD = (domain, action) => {
+  return domain
+    .push(Immutable.Map({
+      done: false,
+      id: _.uniqueId(),
+      name: action.data.name
+    }));
 };
 
 /**
  * @param {Immutable.List} domain
  * @param {Object} action
- * @param {String} action.data.name
- * @return {Immutable.List}
+ * @param {number} action.data.id
+ * @returns {Immutable.List}
  */
-TASK_ADD = (domain, action) => {
-    return domain
-        .push(Immutable.Map({
-            id: _.uniqueId(),
-            name: action.data.name,
-            done: false
-        }));
+const TASK_DONE = (domain, action) => {
+  const index = domain.findIndex((item) => {
+    return item.get('id') === action.data.id;
+  });
+
+  return domain
+    .update(index, (task) => {
+      return task.set('done', true);
+    });
 };
 
 /**
  * @param {Immutable.List} domain
  * @param {Object} action
- * @param {Number} action.data.id
- * @return {Immutable.List}
+ * @param {number} action.data.id
+ * @returns {Immutable.List}
  */
-TASK_DONE = (domain, action) => {
-    let index;
+const TASK_UNDONE = (domain, action) => {
+  const index = domain
+    .findIndex((item) => {
+      return item.get('id') === action.data.id;
+    });
 
-    index = domain.findIndex((item) => item.get('id') === action.data.id);
-
-    return domain
-        .update(index, (task) => {
-            return task.set('done', true);
-        });
+  return domain
+    .update(index, (task) => {
+      return task.set('done', false);
+    });
 };
 
-/**
- * @param {Immutable.List} domain
- * @param {Object} action
- * @param {Number} action.data.id
- * @return {Immutable.List}
- */
-TASK_UNDONE = (domain, action) => {
-    let index;
-
-    index = domain
-        .findIndex((item) => item.get('id') === action.data.id);
-
-    return domain
-        .update(index, (task) => {
-            return task.set('done', false);
-        });
-};
-
-export {
-    CONSTRUCT,
-    TASK_ADD,
-    TASK_DONE,
-    TASK_UNDONE
-};
+export default createReducer(initialState, {
+  TASK_ADD,
+  TASK_DONE,
+  TASK_UNDONE
+});
